+++ src/rest.ts
@@ -6,5 +6,5 @@ export {HTTP_HEADER_ACCEPT, HTTP_HEADER_CONTENT_TYPE} from './rest/constants';
 export {Body, BaseUrl, DefaultHeaders, Header, Headers, JsonContentType, ParameterTransform, Path, Produces, Query, QueryObject, ResponseTransform, FullHttpResponse, ReportProgress, DisableInterceptor, ProgressIndicatorGroup, AcceptAny, Cache} from './rest/decorators';
 export {REST_METHOD_MIDDLEWARES, REST_MIDDLEWARES_ORDER} from './rest/tokens';
 export {BASIC_DEFAULT_REST_METHOD_MIDDLEWARES, BASIC_DEFAULT_REST_MIDDLEWARES_ORDER} from './rest/defaults';
-export {ReportProgressMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, ResponseTypeMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, HeaderParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeadersMiddleware, CacheMiddleware} from './rest/middlewares';
+export {ReportProgressMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, ResponseTypeMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, HeaderParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeadersMiddleware, CacheMiddleware, LoggerMiddleware} from './rest/middlewares';
 export {buildMiddlewares, getType, isNotType, not} from './rest/utils';