+++ src/rest/defaults.ts
@@ -1,7 +1,7 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware, HeadersMiddleware, CacheMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware, HeadersMiddleware, CacheMiddleware, LoggerMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
@@ -15,6 +15,7 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
     HeadersMiddleware,
     HeaderParameterMiddleware,
     ProducesMiddleware,
+    LoggerMiddleware,
     AdditionalDataMiddleware,
     IgnoredInterceptorsMiddleware,
     ResponseTransformMiddleware,
@@ -28,6 +29,7 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
  */
 export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
 [
+    LoggerMiddleware,
     ResponseTypeMiddleware,
     ReportProgressMiddleware
 ];